migrate liftArray out of tsGenerics
